add elements for pokemonListTest component and remove unnecessary files
